backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
django-import-export==4.3.9
django-location-field==2.7.3
django-ninja==1.4.3
django-simplemde==0.1.4
django-prometheus==2.4.1
django-redis==6.0.0
django-unfold==0.63.0
dnspython==2.7.0
email_validator==2.2.0
flower==2.0.1
frozenlist==1.7.0
gunicorn==23.0.0
http_ece==1.2.1
humanize==4.12.3
idna==3.10
kombu==5.5.4
Markdown==3.8.2
multidict==6.6.4
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
py-vapid==1.9.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
pytz==2025.2
pywebpush==2.0.3
redis==6.4.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
jdatetime
